+++ collect.rb
@@ -1,78 +1,50 @@
-# coding: utf-8
-require 'logger'
 require 'nokogiri'
 require_relative 'config/initialize'
-require_relative 'lib/http_client'
+require_relative 'lib/collect_util'
+require_relative 'lib/denebola_logger'
 
 BACKUP_DIR = File.join(APPLICATION_ROOT, 'backup')
-
-logger = Logger.new('log/collect.log')
-logger.formatter = proc do |severity, datetime, progname, message|
-  time = datetime.utc.strftime(Settings.logger.time_format)
-  log = "[#{severity}] [#{time}]: #{message}"
-  puts log if ENV['STDOUT'] == 'on'
-  "#{log}\n"
-end
+logger = DenebolaLogger.new(Settings.logger.path.collect)
+CollectUtil.logger = logger
 
 begin
   from = ARGV.find {|arg| arg.start_with?('--from=') }
-  from = from ? Date.parse(from.match(/\A--from=(.*)$\z/)[1]) : (Date.today - 7)
+  from = from ? Date.parse(from.match(/\A--from=(.*)$\z/)[1]) : (Date.today - 30)
   to = ARGV.find {|arg| arg.start_with?('--to=') }
   to = to ? Date.parse(to.match(/\A--to=(.*)\z/)[1]) : Date.today
-rescue Exception => e
+rescue ArgumentError => e
   logger.error(e.backtrace.join("\n"))
   raise e
 end
 
-client = HTTPClient.new
-
-Settings.backup_dir.to_h.values.each {|path| FileUtils.mkdir_p(File.join(BACKUP_DIR, path)) }
+Settings.backup_dir.to_h.values.each do |path|
+  FileUtils.mkdir_p(File.join(BACKUP_DIR, path))
+  removed_files = Dir[File.join(BACKUP_DIR, path, '*')].select do |file_path|
+    File.zero?(file_path)
+  end
+  FileUtils.rm(removed_files) if removed_files
+end
 
 (from..to).each do |date|
   date = date.strftime('%Y%m%d')
 
   file_path = File.join(BACKUP_DIR, Settings.backup_dir.race_list, "#{date}.txt")
-  race_ids = if File.exists?(file_path)
-               ids = File.read(file_path).split("\n")
-               logger.info(:file_path => File.basename(file_path), :race_ids => ids)
-               ids
-             else
-               res = client.get("#{Settings.url}#{Settings.path.race_list}/#{date}")
-               ids = res.body.scan(%r[.*/race/(\d+)]).flatten
-               logger.info(:uri => res.uri.to_s, :status => res.code, :race_ids => ids)
-               File.open(file_path, 'w') {|out| out.write(ids.join("\n")) }
-               ids
-             end
+  race_ids = CollectUtil.get_race_ids(file_path, date)
 
   race_ids.each do |race_id|
     file_path = File.join(BACKUP_DIR, Settings.backup_dir.race, "#{race_id}.html")
-    html = if File.exists?(file_path)
-             html = File.read(file_path)
-             logger.info(:resource => 'race', :file_path => File.basename(file_path))
-             html
-           else
-             res = client.get("#{Settings.url}#{Settings.path.race}/#{race_id}")
-             logger.info(:resource => 'race', :uri => res.uri.to_s, :status => res.code)
-             html = res.body.encode('utf-8', 'euc-jp', :undef => :replace, :replace => '?')
-             html.gsub!('&nbsp;', ' ')
-             File.open(file_path, 'w') {|out| out.write(html) }
-             html
-           end
+    html = CollectUtil.get_race_html(file_path, race_id)
 
     parsed_html = Nokogiri::HTML.parse(html)
-    _, *entries = parsed_html.xpath('//table[contains(@class, "race_table")]').search('tr')
+    _, *entries =
+      parsed_html.xpath('//table[contains(@class, "race_table")]').search('tr')
 
     entries.each do |entry|
       horse_link = entry.search('td')[3].first_element_child.attribute('href').value
-      horse_id = horse_link.match(/\/horse\/(?<horse_id>\d+)\/?/)[:horse_id]
+      horse_id = horse_link.match(%r{/horse/(?<horse_id>\d+)/?})[:horse_id]
 
       file_path = File.join(BACKUP_DIR, Settings.backup_dir.horse, "#{horse_id}.html")
-      unless File.exists?(file_path)
-        res = client.get("#{Settings.url}#{Settings.path.horse}/#{horse_id}")
-        logger.info(:resource => 'horse', :uri => res.uri.to_s, :status => res.code)
-        html = res.body.encode('utf-8', 'euc-jp', :undef => :replace, :replace => '?')
-        File.open(file_path, 'w') {|out| out.write(html.gsub('&nbsp;', ' ')) }
-      end
+      CollectUtil.get_horse_html(file_path, horse_id)
     end
   end
 end