+++ db/connect.rb
@@ -1,5 +1,10 @@
 require 'active_record'
 require 'mysql2'
 require_relative '../config/initialize'
+require_relative '../lib/denebola_logger'
 
-ActiveRecord::Base.establish_connection(Settings.mysql.to_h)
+ActiveRecord::Base.logger = DenebolaLogger.new(Settings.logger.path.database)
+ActiveRecord::Base.logger.level = Settings.logger.level.to_sym
+settings = Settings.mysql.map {|key, value| [key.to_s, value] }.to_h
+settings['database'] = Settings.mysql.database[ENV['RAILS_ENV']]
+ActiveRecord::Base.establish_connection(settings)