+++ db/migrate/002_create_entries.rb
@@ -1,15 +1,16 @@
 class CreateEntries < ActiveRecord::Migration[4.2]
   def change
     create_table :entries do |t|
-      t.integer :number, :null => false
-      t.integer :age, :null => false
-      t.float :burden_weight, :null => false
+      t.integer :age, null: false
+      t.float :burden_weight, null: false
+      t.integer :number, null: false
       t.float :weight
-      t.references :race
 
-      t.timestamps :null => false
+      t.references :race, after: :weight
+
+      t.timestamps null: false
     end
 
-    add_index :entries, [:race_id, :number], :unique => true
+    add_index :entries, %i[race_id number], unique: true
   end
 end