+++ db/migrate/003_create_results.rb
@@ -1,13 +1,14 @@
 class CreateResults < ActiveRecord::Migration[4.2]
   def change
     create_table :results do |t|
-      t.string :order, :null => false
-      t.references :race
-      t.references :entry
+      t.string :order, null: false
 
-      t.timestamps :null => false
+      t.references :entry
+      t.references :race
+
+      t.timestamps null: false
     end
 
-    add_index :results, [:race_id, :entry_id], :unique => true
+    add_index :results, %i[race_id entry_id], unique: true
   end
 end