+++ db/migrate/005_add_grade_to_races.rb
@@ -1,5 +1,5 @@
 class AddGradeToRaces < ActiveRecord::Migration[4.2]
   def change
-    add_column :races, :grade, :string, :after => :round
+    add_column :races, :grade, :string, after: :distance
   end
 end