+++ db/migrate/010_create_horses.rb
@@ -1,10 +1,11 @@
 class CreateHorses < ActiveRecord::Migration[4.2]
   def change
     create_table :horses do |t|
-      t.string :horse_id, :null => false
-      t.timestamps :null => false
+      t.string :horse_id, null: false
+
+      t.timestamps null: false
     end
 
-    add_index :horses, [:horse_id], :unique => true
+    add_index :horses, %i[horse_id], unique: true
   end
 end