+++ db/migrate/012_add_race_id_to_races.rb
@@ -1,0 +1,5 @@
+class AddRaceIdToRaces < ActiveRecord::Migration[4.2]
+  def change
+    add_column :races, :race_id, :string, null: false, after: :place
+  end
+end