+++ db/migrate/013_add_last_race_order_to_horses_and_features.rb
@@ -1,0 +1,6 @@
+class AddLastRaceOrderToHorsesAndFeatures < ActiveRecord::Migration[4.2]
+  def change
+    add_column :horses, :last_race_order, :integer, after: :id
+    add_column :features, :last_race_order, :integer, after: :jockey
+  end
+end