+++ db/migrate/019_add_running_style_to_horses.rb
@@ -1,0 +1,5 @@
+class AddRunningStyleToHorses < ActiveRecord::Migration[4.2]
+  def change
+    add_column :horses, :running_style, :string, after: :horse_id
+  end
+end