+++ db/migrate/023_change_running_style_on_horses.rb
@@ -1,0 +1,5 @@
+class ChangeRunningStyleOnHorses < ActiveRecord::Migration[4.2]
+  def change
+    change_column :horses, :running_style, :string, null: false
+  end
+end