+++ db/migrate/026_change_direction_on_races.rb
@@ -1,0 +1,7 @@
+# coding: utf-8
+
+class ChangeDirectionOnRaces < ActiveRecord::Migration[4.2]
+  def change
+    change_column_null :races, :direction, false, '障'
+  end
+end