+++ spec/factories/horse.rb
@@ -1,0 +1,8 @@
+# coding: utf-8
+
+FactoryBot.define do
+  factory :horse do
+    horse_id { '1' * 8 }
+    running_style { '逃げ' }
+  end
+end